selinux_labels test: work with podman
First, per dwalsh, use '--security-opt=...' instead of '...opt:...',
i.e. equals sign instead of colon. The colon form is deprecated
in docker and not supported in podman.

Second, skip tests that check for dockerd or containerd when
testing podman; we detect by adding a 'podman-V-R' string to
the 'version' output in our filter